fastapi-backend/pyproject.toml
[tool.poetry]
name = "fin-path-insight-fastapi"
version = "0.1.0"
description = "FastAPI backend for FinPath Insight"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.104.1"
uvicorn = "0.24.0"
gunicorn = "21.2.0"
pydantic = "2.4.2"
pydantic-settings = "2.0.3"
python-dotenv = "1.0.0"
httpx = "0.25.1"
pinecone-client = "2.2.4"
google-generativeai = "0.3.1"
python-multipart = "0.0.6"
PyPDF2 = "3.0.1"
pandas = "2.1.1"
openpyxl = "3.1.2"
numpy = "1.24.3"
python-jose = "3.3.0"
passlib = "1.7.4"
bcrypt = "4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
